Move models from Market api entry to models dedicated package
